fix: fixing global dependency management for project packages by using YARN
From now on, Yarn needs to be used as a package manager, solely for the reason that yarn workspaces
make it trivial to manage all the package's dependencies of a cantara project, which was one of the
main bug sources until now. Webpack aliases just weren't enough. This is considered a bbreaking
change because all consumers expect to have yarn installed globally now. This also holds true for CI
environments. Additionaly, the 'appDependencyAliases' concept was removed. Moving to yarn should
also makestartup times of development environments notably faster.

BREAKING CHANGE: All consumers are expected to have the Yarn Package Manager installed globally.